Add files via upload
This program creates classes and subclasses with their corresponding data fields and constructor ,call the super class constructor on each class to inherit the data field "name" of the super class, override the toString() method in each class to display the class name and person's name. Then writes a test program in the main class and invokes the to.String() methods of each class.